+++ src/amp_simulate/launch/robot_launch.py
@@ -17,7 +17,6 @@
 """Launch Webots Tesla driver."""
 
 import os
-import pathlib
 import launch
 from launch.substitutions import LaunchConfiguration
 from launch.actions import DeclareLaunchArgument
@@ -26,34 +25,7 @@ from launch import LaunchDescription
 from ament_index_python.packages import get_package_share_directory
 from launch_ros.actions import Node
 from webots_ros2_driver.webots_launcher import WebotsLauncher
-from webots_ros2_driver.utils import controller_url_prefix
-
-
-def get_ros2_nodes(*args):
-    package_dir = get_package_share_directory('amp_simulate')
-    robot_description = pathlib.Path(os.path.join(package_dir, 'resource', 'tesla_webots.urdf')).read_text()
-
-    print('WEBOTS_CONTROLLER_URL', controller_url_prefix(), 'vehicle')
-
-    tesla_driver = Node(
-        package='webots_ros2_driver',
-        executable='driver',
-        output='screen',
-        additional_env={'WEBOTS_CONTROLLER_URL': controller_url_prefix() + 'vehicle'},
-        parameters=[
-            {'robot_description': robot_description},
-        ]
-    )
-
-    lane_follower = Node(
-        package='amp_simulate',
-        executable='lane_follower',
-    )
-
-    return [
-        # lane_follower,
-        tesla_driver,
-    ]
+from webots_ros2_driver.webots_controller import WebotsController
 
 
 def generate_launch_description():
@@ -65,13 +37,17 @@ def generate_launch_description():
         ros2_supervisor=True
     )
 
-    # The following line is important!
-    # This event handler respawns the ROS 2 nodes on simulation reset (supervisor process ends).
-    reset_handler = launch.actions.RegisterEventHandler(
-        event_handler=launch.event_handlers.OnProcessExit(
-            target_action=webots._supervisor,
-            on_exit=get_ros2_nodes,
-        )
+    robot_description_path = os.path.join(package_dir, 'resource', 'tesla_webots.urdf')
+    kart_driver = WebotsController(
+        robot_name='vehicle',
+        parameters=[
+            {'robot_description': robot_description_path}
+        ],
+        respawn=True
+    )
+    lane_follower = Node(
+        package='amp_lane',
+        executable='lane_follower',
     )
 
     return LaunchDescription([
@@ -82,20 +58,16 @@ def generate_launch_description():
         ),
         webots,
         webots._supervisor,
+        kart_driver,
+        lane_follower,
 
         # This action will kill all nodes once the Webots simulation has exited
         launch.actions.RegisterEventHandler(
             event_handler=launch.event_handlers.OnProcessExit(
                 target_action=webots,
                 on_exit=[
-                    launch.actions.UnregisterEventHandler(
-                        event_handler=reset_handler.event_handler
-                    ),
                     launch.actions.EmitEvent(event=launch.events.Shutdown())
                 ],
             )
-        ),
-
-        # Add the reset event handler
-        reset_handler
-    ] + get_ros2_nodes())
+        )
+    ])