+++ src/amp_simulate/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package amp_simulate
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2023.1.0 (2023-06-29)
+------------------
+* Clean simulation reset in launch file.
+* Update driver node to new WebotsController node.
+
 2023.0.2 (2023-02-07)
 ------------------
 * Updated supervisor launch.